Rename js files to ts and fix imports
